update to logo
You might notice that this reverts the embedding of the logo in the page, the original idea was to improve loading by preventing another http request, but I discovered that even though the logo was embedded in the html, it was still being loaded again for the favico. So I might as well leave it external.