remove unused packageJson
When I originally started this project I installed modelviewer through node.
Because the site is hosted on GitHub pages, the model viewer JS was added to the JS directory. Node isn't currently being used for anything else on the site, so I've removing these outdated files to better reflect the state of the project.